backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
ansible-runner==2.4.0
PyYAML==6.0.1
openpyxl==3.1.2
pandas==2.1.1
python-dotenv==1.0.0
gunicorn==21.2.0
Flask-Compress==1.13
psutil==5.9.5
# Additional libraries for enhanced functionality
numpy==1.24.3
xlrd==2.0.1
xlsxwriter==3.1.2
# For better logging and monitoring
structlog==23.1.0
# For better error handling
tenacity==8.2.2
# Database and migrations
psycopg2-binary==2.9.9
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
# Authentication and security
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.3
Flask-Admin==1.6.1
APScheduler==3.10.4
# API development
Flask-RESTful==0.3.10
marshmallow==3.20.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
Flask-Limiter==3.5.0
redis==4.6.0